services/vessel-economics/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database drivers
psycopg2-binary==2.9.9
cassandra-driver==3.29.0

# Kafka
kafka-python==2.0.2

# Monitoring and metrics
prometheus-client==0.19.0

# Data processing
pandas==2.1.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Logging
python-json-logger==2.0.7